Separated conn process from conn proxy
Created a factory to get the conn process
Removed conn process from conn proxy
Updated DefaultTestSelector to use the new factory
