+++ mbed_host_tests/host_tests_event_loop/base_event_loop.py
@@ -18,8 +18,7 @@ limitations under the License.
 
 import traceback
 
-from Queue import Empty as QueueEmpty   # Queue here refers to the module, not a class
-
+from Queue import Empty as QueueEmpty
 from time import time
 from abc import ABCMeta, abstractmethod
