+++ mbed_host_tests/host_tests_event_loop/default_event_loop.py
@@ -20,7 +20,6 @@ import traceback
 from time import time
 from mbed_host_tests import BaseHostTest
 from mbed_host_tests import get_host_test
-from mbed_host_tests.host_tests_conn_proxy import conn_process
 from mbed_host_tests.host_tests_event_loop import BaseEventLoop
 
 class DefaultEventLoop(BaseEventLoop):
